Update VRC7 emulator core (emu2413)

+++ Source/APU/nsfplay/xgm/devices/Sound/legacy/emu2413.h
@@ -102,9 +102,9 @@ typedef struct __OPLL {
 
   uint32_t adr;
 
-  uint32_t inp_step;
-  uint32_t out_step;
-  uint32_t out_time;
+  double inp_step;
+  double out_step;
+  double out_time;
 
   uint8_t reg[0x40];
   uint8_t test_flag;
